extract out resolveSchemaChain to own file
